Refactor GitHub client auth and dependency imports
I've added the octokit enterprise plugin. This will help us include the enterprise endpoints that are not
included by default. Including this plugin required that we change the way we are authenticating and creating
the Octokit client instance. We are now using an instance with the pluging added to create our client and it's auth.

I've also removed the import of octokit/graphql. This module is already included in the Octokit/rest library that we
are using. Removing the use of the octokit/graphql module let us also remove the function that creates a graphql client.

All the client creation code we need is now in the buildGitHubClient function. The client returned by this funciton is
now passed into the data export functions for API v3 and v4.