Refactor v3 data endpoints
I've refactored the requestV3Entries function to include the audit log source argument and the enterprise
audit log endpoints. The source argument will determine what endpoint to use.

I've also renamed the org argument to entity. Since we are now working with orgs or enterprises this name seemed
better suited.

The clients passed to the data request functions have also been changed to be the same name. We are now using the
graphql client that is included in the Octokit/rest package.